# wp-content/plugins/gym-core/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gym Core REST API",
    "description": "REST API for Haanpaa Martial Arts gym management. All endpoints use the gym/v1 namespace and return a standard envelope: { success: true, data: <payload>, meta?: { pagination? } }.",
    "version": "2.1.0",
    "contact": {
      "name": "Andrew Wikel"
    }
  },
  "servers": [
    {
      "url": "/wp-json/gym/v1",
      "description": "WordPress REST API (gym/v1 namespace)"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WP-Nonce",
        "description": "WordPress REST nonce (cookie authentication)"
      },
      "applicationPassword": {
        "type": "http",
        "scheme": "basic",
        "description": "WordPress Application Password"
      }
    },
...
  },
  "paths": {
    "/check-in": {
      "post": {
        "summary": "Record member check-in",
        "tags": ["Attendance"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "class_id", "method"],
                "properties": {
                  "user_id": { "type": "integer" },
                  "class_id": { "type": "integer" },
                  "method": { "type": "string", "enum": ["qr_scan", "member_id", "name_search", "manual"] },
                  "location": { "type": "string", "description": "Location slug fallback for Open Mat" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Check-in recorded" },
          "401": { "description": "Not authenticated" },
          "403": { "description": "Requires gym_check_in_member or manage_options" }
        }
      }
    },
...
    "/attendance/today": {
      "get": {
        "summary": "Get today's check-ins",
        "tags": ["Attendance"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "parameters": [
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "class_id", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Today's check-in records" },
          "403": { "description": "Requires gym_view_attendance or manage_woocommerce" }
        }
      }
    },
...
  },
...
}